add error, add interceptor

+++ src/products/products.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Param, Delete, Post, Body, ParseUUIDPipe, Put } from '@nestjs/common';
+import { Controller, Get, Param, Delete, Post, Body, ParseUUIDPipe, Put, NotFoundException } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
 import { UpdateProductDTO } from './dtos/update-product.dto';
@@ -12,12 +12,16 @@ export class ProductsController {
   }
 
   @Get('/:id')
-  public getById(@Param('id', new ParseUUIDPipe()) id: string) {
-    return this.productsService.getById(id);
+  getById(@Param('id', new ParseUUIDPipe()) id: string) {
+    const prod = this.productsService.getById(id);
+    if (!prod) throw new NotFoundException('Product not found');
+    return prod;
   }
 
   @Delete('/:id')
   deleteById(@Param('id', new ParseUUIDPipe()) id: string){
+    if(!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
     this.productsService.deleteByID(id);
     return { success: true};
   }
@@ -29,8 +33,12 @@ export class ProductsController {
 
   @Put('/:id')
   update(
-    @Param('id', new ParseUUIDPipe()) id: string, 
-    @Body() productData: UpdateProductDTO) {
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
     this.productsService.updateById(id, productData);
     return { success: true };
   }
